backend/backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-background-tasks==1.2.8
django-cors-headers==4.6.0
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
exceptiongroup==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
mysql-connector-python==9.3.0
packaging==24.2
pillow==11.0.0
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
PyJWT==2.10.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-django==4.11.1
pytest-playwright==0.7.0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
ruff==0.11.5
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
